Creates pages folder in Settings and move webhooks in it
