+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/EditView/components/utils/fieldsRegex.js
