+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/EditView/index.js
@@ -17,7 +17,7 @@ import React, { useCallback, useMemo } from 'react';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useHistory, useRouteMatch } from 'react-router-dom';
 import WebhookForm from './components/WebhookForm';
-import { useModels } from '../../../hooks';
+import { useModels } from '../../../../../hooks';
 import cleanData from './utils/formatData';
 
 const EditView = () => {
