+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/EditView/utils/formatData.js
