+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/ProtectedCreateView/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CheckPagePermissions } from '@strapi/helper-plugin';
-import adminPermissions from '../../../permissions';
+import adminPermissions from '../../../../../permissions';
 import EditView from '../EditView';
 
 const ProtectedCreateView = () => (
